+++ src/Controller/AuthenticationController.php
@@ -152,120 +152,6 @@ class AuthenticationController extends AbstractController
     }
 
     /**
-     * @throws InvalidArgumentException
-     *
-     * Requires session cookie to be set to a valid session.
-     */
-    #[Route(path: '/authentication/status', name: 'app_identity_authentication_status', methods: ['GET'])]
-    public function authenticationStatus(): JsonResponse
-    {
-        $nameId = $this->authenticationService->getNameId();
-        $sari = $this->stateHandler->getRequestId();
-        $logger = WithContextLogger::from($this->logger, ['nameId' => $nameId, 'sari' => $sari]);
-        $logger->info('Request for authentication status');
-
-        // Do we have a valid GSSP authentication AuthnRequest in this session?
-        if (!$this->authenticationService->authenticationRequired()) {
-            $logger->error('There is no pending authentication request from SP');
-            return $this->refreshAuthenticationPage();
-        }
-
-        // Check if the user is authenticated. This check will work if the user has completed the authentication
-        // within the last Tiqr_Service::LOGIN_EXPIRE seconds (1 hour).
-        // Uses our SessionId.
-        $isAuthenticated = $this->tiqrService->isAuthenticated();
-
-        if ($isAuthenticated) {
-            $logger->info('Send json response is authenticated');
-
-            return $this->refreshAuthenticationPage();
-        }
-
-        if ($this->authenticationChallengeIsExpired()) {
-            return $this->timeoutNeedsManualRetry();
-        }
-
-        $logger->info('Send json response is not authenticated');
-
-        return $this->scheduleNextPollOnAuthenticationPage();
-    }
-
-    /**
-     * Check if the authentication challenge is expired.
-     *
-     * If the challenge is expired, the page should be refreshed so a new
-     * challenge and QR code is generated.
-     *
-     * @return bool
-     */
-    private function authenticationChallengeIsExpired(): bool
-    {
-        // The use of authenticationUrl() here is a hack, because it depends on an implementation detail
-        // of this function.
-        // Effectively this does a $this->_stateStorage->getValue(self::PREFIX_CHALLENGE . $sessionKey);
-        // To check that the session key still exists in the Tiqr_Service's state storage
-
-        // TODO: Refactor this so it does not depend on the implementation of the TiqrService
-        //       i.e. we should keep track of authentication timeout ourselves
-        //       Uses this method creates noise in the logs because it logs errors because keys are not found
-        //       in the StateStorage. But this is not an error here, it is expected bacause expired keys
-        //       cannot be found in the statestorage by definition.
-        //       If you were truly interested in the authenticationUrl(), then logging the errors is correct.
-        try {
-            $this->tiqrService->authenticationUrl();
-        } catch (Exception) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Authentication is pending, schedule a new poll action.
-     *
-     * @return JsonResponse
-     */
-    private function scheduleNextPollOnAuthenticationPage(): JsonResponse
-    {
-        return $this->generateAuthenticationStatusResponse('pending');
-    }
-
-    /**
-     * Generate a response for authentication.html: Ask the user to retry.
-     *
-     * @return JsonResponse
-     */
-    private function timeoutNeedsManualRetry(): JsonResponse
-    {
-        return $this->generateAuthenticationStatusResponse('challenge-expired');
-    }
-
-    /**
-     * Generate a response for authentication.html: refresh the page.
-     *
-     * @return JsonResponse
-     */
-    private function refreshAuthenticationPage(): JsonResponse
-    {
-        return $this->generateAuthenticationStatusResponse('needs-refresh');
-    }
-
-    /**
-     * Generate a status response for authentication.html.
-     *
-     * The javascript in the authentication page expects one of three statuses:
-     *
-     *  - pending: waiting for user action, schedule next poll
-     *  - needs-refresh: refresh the page (the /authentication page will handle the success or error)
-     *  - challenge-expired: Message user challenge is expired, let the user give the option to retry.
-     *
-     * @return JsonResponse
-     */
-    private function generateAuthenticationStatusResponse(string $status): JsonResponse
-    {
-        return new JsonResponse($status);
-    }
-
-    /**
      * Generate a notification response for authentication.html.
      *
      * The javascript in the authentication page expects one of three statuses:
